moving to float SYMMATRIX type and fixing no-compile problems
git-svn-id: https://advants.svn.sourceforge.net/svnroot/advants@264 fd2da0a4-d968-48df-9cb5-d8c496287cbf

+++ Tensor/itkPreservationOfPrincipalDirectionTensorReorientationImageFilter.cxx
@@ -170,9 +170,9 @@ dest[2]=v1[2]-v2[2];}
 
 
 template<typename TTensorImage, typename TVectorImage>
- itk::SymmetricSecondRankTensor< double, 3 >
+ itk::SymmetricSecondRankTensor< float, 3 >
 PreservationOfPrincipalDirectionTensorReorientationImageFilter<TTensorImage,TVectorImage>
-::ReorientTensors( itk::SymmetricSecondRankTensor< double, 3 > fixedTens, itk::SymmetricSecondRankTensor< double, 3 > movingTens,  typename TTensorImage::IndexType index)
+::ReorientTensors( itk::SymmetricSecondRankTensor< float, 3 > fixedTens, itk::SymmetricSecondRankTensor< float, 3 > movingTens,  typename TTensorImage::IndexType index)
 {
 // get from and to vectors
