+++ Tensor/itkPreservationOfPrincipalDirectionTensorReorientationImageFilter.h
@@ -64,7 +64,7 @@ public:
   typedef itk::Image<float,ImageDimension> FloatImageType;
 
   //  typedef Vector<float, 6> TensorType;
-  typedef itk::SymmetricSecondRankTensor< double, 3 >  TensorType; 
+  typedef itk::SymmetricSecondRankTensor< float, 3 >  TensorType; 
   typedef Image<TensorType, ImageDimension> TensorImageType;
   typedef typename TensorImageType::Pointer TensorImagePointer;
   typedef vnl_matrix<float> VnlMatrixType;
